overall code + performance improvements
